Move syntax APIs to separate namespace/folder
