+++ Vice/vc/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,4 +1,4 @@
-﻿namespace Vice.CodeAnalysis
+﻿namespace Vice.CodeAnalysis.Syntax
 {
     public enum SyntaxKind
     {
