+++ Vice/vc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Vice.CodeAnalysis
+namespace Vice.CodeAnalysis.Syntax
 {
     public abstract class SyntaxNode
     {
