+++ Vice/vc/CodeAnalysis/Syntax/SyntaxTree.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace Vice.CodeAnalysis
+namespace Vice.CodeAnalysis.Syntax
 {
     public sealed class SyntaxTree
     {
