+++ Vice/vc/CodeAnalysis/Syntax/UnaryExpressionSyntax.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Vice.CodeAnalysis
+namespace Vice.CodeAnalysis.Syntax
 {
     public sealed class UnaryExpressionSyntax : ExpressionSyntax
     {
